Removing long (part 4): Remove support for serializing longs

+++ src/item_group.h
@@ -152,7 +152,7 @@ class Item_modifier
          * Charges to spawn the item with, if this turns out to
          * be negative, the default charges are used.
          */
-        std::pair<long, long> charges;
+        std::pair<int, int> charges;
         /**
          * Ammo for guns. If NULL the gun spawns without ammo.
          * This takes @ref charges and @ref with_ammo into account.
